+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 `StretchScrollView` adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0](https://github.com/APUtils/StretchScrollView/releases/tag/1.1.0)
+Released on 07/20/2017.
+
+#### Added
+- Now scrollView pays attention to top contentInset so translucent navigation bar won't overlap image.
+  - Added by [Anton Plebanovich](https://github.com/anton-plebanovich).
+
 ## [1.0.1](https://github.com/APUtils/StretchScrollView/releases/tag/1.0.1)
 Released on 07/20/2017.